requirements.txt
configuration-registry
consul-service-discovery
click
